refactor: extract decode attention A2F scheduling
